Rename folder + adjust namespace
