+++ src/Hackathon.Fiap.UseCases/Patients/GetByUserId/GetPatientByUserIdQuery.cs
@@ -1,2 +1,3 @@
-﻿namespace Hackathon.Fiap.UseCases.Patients.GetByUser;
+﻿namespace Hackathon.Fiap.UseCases.Patients.GetByUserId;
+
 public record GetPatientByUserIdQuery(string UserId) : IQuery<PatientDto?>;
